requirements.txt
aiohttp==3.10.5
aiosignal==1.3.2
aiounittest==1.3.0
anyio==4.7.0
applicationinsights==0.11.9
attrs==24.3.0
azure-cognitiveservices-language-luis==0.2.0
azure-core==1.32.0
azure-identity==1.19.0
azure-storage-blob==12.7.0
azure-storage-queue==12.4.0
azure-cosmos==4.7.0
blis==1.1.0
botbuilder-ai==4.16.2
botbuilder-applicationinsights==4.16.2
botbuilder-azure==4.16.2
botbuilder-core==4.16.2
botbuilder-dialogs==4.16.2
botbuilder-integration-aiohttp==4.16.2
botbuilder-schema==4.16.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
confection==0.1.5
coverage==7.2.1
cryptography==41.0.3
cymem==2.0.10
distro==1.9.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib-metadata==6.0.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
jsonpickle==1.4.2
langcodes==3.5.0
marisa-trie==1.2.1
MarkupSafe==3.0.2
mpmath==1.3.0
msal==1.31.1
msal-extensions==1.2.0
msrest==0.7.1
msrestazure==0.6.4.post1
multidict==6.1.0
murmurhash==1.0.11
networkx==3.4.2
nltk==3.8.1
numpy==2.2.1
oauthlib==3.2.2
openai==1.58.1
packaging==24.1
pillow==11.0.0
portalocker==2.10.1
preshed==3.0.9
propcache==0.2.1
psutil==6.1.1
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytest==7.4.0
pytz==2024.2
pywin32==308
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
scipy==1.14.1
sentencepiece==0.2.0
setuptools==75.6.0
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
spacy==3.8.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.0
sympy==1.13.1
tenacity==8.2.2
textblob==0.17.1
thinc==8.3.3
tokenizers==0.21.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.47.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vaderSentiment==3.3.2
wasabi==1.1.3
Werkzeug==3.1.3
wrapt==1.17.0
yarl==1.18.3
pyodbc==4.0.32
